Create car-parking-sensor-arduino
C++ Program simulating the parking sensors of a vehicle.

Material needed:
Arduino, button, LEDs, Buzzer, LCD display.

How does it work? 
The ultrasound sensor calculates the distance (using the time of the travel of the soundwave) of the objects situated in front of his area of reach, and if it detects something within a radius of 30 cm, the buzzer goes off and the LED lights up. Depending on the distance of the object in question, the frequency of the buzzer will vary (the closer the object is, the faster the buzzer emits the sound, as will the LED, which flashes according to the speed of the buzzer). The LCD screen displays the distance measured by the sensor. This program was made using the distance of 30cm to facilitate demonstrations but it can easily serve for other distances aswell.